requirements.txt
Django==1.5.4
djangorestframework==2.3.8
GitPython==0.3.2.RC1
